feat(liikunta-374): remove venue from sports and use it from liikunta federeation only
