--- apps/sports-helsinki/src/domain/graphql/paginationSchema.ts
@@ -1,13 +1,0 @@
-import gql from 'graphql-tag';
-
-const typeDefs = gql`
-  type PageInfo {
-    hasPreviousPage: Boolean!
-    hasNextPage: Boolean!
-    startCursor: String
-    endCursor: String
-    count: Int!
-  }
-`;
-
-export default typeDefs;
