--- apps/sports-helsinki/src/domain/graphql/venue/instructions/VenueEnricher.ts
@@ -1,5 +1,0 @@
-import type { AnyObject, Context } from '../../../nextApi/types';
-
-export default interface VenueEnricher<I = AnyObject, O = AnyObject> {
-  getEnrichments(data: I, context: Context): Promise<Partial<O>>;
-}
