--- apps/sports-helsinki/src/domain/graphql/venue/parseVenueId.ts
@@ -1,35 +1,0 @@
-import { Sources } from '../../app/appConstants';
-import type { Source } from '../../nextApi/types';
-
-const SUPPORTED_SOURCES: string[] = [Sources.TPREK, Sources.LINKED];
-
-export class IdParseError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'IdParseError';
-  }
-}
-
-export default function parseVenueId(
-  venueId: string | string[]
-): [Source, string] | null {
-  if (Array.isArray(venueId)) {
-    throw Error('Array IDs are not supported');
-  }
-
-  const [source, id] = venueId.split(':');
-
-  if (!source) {
-    throw new IdParseError('Could not find source from venue id');
-  }
-
-  if (!SUPPORTED_SOURCES.includes(source)) {
-    throw new IdParseError(`Found unsupported source from venue id: ${source}`);
-  }
-
-  if (!id) {
-    throw new IdParseError('Could not find id from venue id');
-  }
-
-  return [source as Source, id];
-}
