--- apps/sports-helsinki/src/domain/nextApi/listVenueFragment.ts
@@ -1,2 +1,0 @@
-export { ListVenueFragmentDoc } from './graphql/__imported__';
-export type { ListVenueFragment } from './graphql/__imported__';
