--- apps/sports-helsinki/src/domain/nextApi/pageInfoFragment.ts
@@ -1,2 +1,0 @@
-export { PageInfoFragmentDoc } from './graphql/__imported__';
-export type { PageInfoFragment } from './graphql/__imported__';
