+++ apps/sports-helsinki/src/types.ts
@@ -1,3 +1,4 @@
+import type { UnifiedSearchVenue } from 'events-helsinki-components';
 import type { SEARCH_ROUTES } from './constants';
 
 export type Language = 'en' | 'fi' | 'sv';
@@ -10,3 +11,16 @@ export enum SUPPORT_LANGUAGES {
 
 // eslint-disable-next-line prettier/prettier
 export type SearchRoute = typeof SEARCH_ROUTES[keyof typeof SEARCH_ROUTES];
+
+export type Node<T> = {
+  cursor: string;
+  node: T;
+};
+
+export type Connection<T> = {
+  edges: Node<T>[];
+};
+
+export type SearchResult = {
+  venue: UnifiedSearchVenue;
+};
