+++ proxies/events-graphql-federation/subgraphs/unified-search/unified-search.graphql
@@ -505,7 +505,7 @@ type NodeMeta @cacheControl(inheritMaxAge: true) {
 
 """
 TODO: merge all free tags, categories, and keywords
-KEYWORDS ARE GIVEN FROM events-proxy (https://tapahtumat-proxy.test.kuva.hel.ninja/proxy/graphql)
+KEYWORDS ARE GIVEN FROM events-proxy (https://events-graphql-proxy.test.hel.ninja/proxy/graphql)
 """
 type KeywordString {
   name: String!
