+++ proxies/venue-graphql-proxy/.env.example
@@ -1,7 +1,6 @@
 NODE_ENV=development
 GRAPHQL_PROXY_DEBUG=1
 GRAPHQL_PROXY_PORT=4200
-GRAPHQL_PROXY_MAP_OPEN_DATA_API_BASE_URL=https://kartta.hel.fi/ws/geoserver/avoindata
 GRAPHQL_PROXY_SENTRY_DSN=
 GRAPHQL_PROXY_SENTRY_ENVIRONMENT=local
